+++ engine/testutil/mock/nodes.go
@@ -102,7 +102,7 @@ type ConsensusNode struct {
 	Receipts        mempool.ExecutionTree
 	Seals           mempool.IncorporatedResultSeals
 	IngestionEngine *consensusingest.Engine
-	MatchingEngine  *matching.Engine
+	MatchingEngine  *matching.Core2
 }
 
 func (cn ConsensusNode) Ready() {
